Pull in part of MetricBuilder from `utilities`
I don't want to rewrite the ConfigDB MQTT client if we're just going to
drop it anyway. But I do need to switch to service-client.
