+++ lib/index.js
@@ -6,6 +6,7 @@ export * from "./well-known.js";
 export * as Interfaces from "./interfaces.js";
 export * as UUIDs from "./uuids.js";
 
+export * from "./sparkplug/metrics.js";
 export * from "./sparkplug/util.js";
 
 /* XXX I don't want to export these; maybe I want a /deps library? */
