+++ lib/sparkplug/util.js
@@ -96,26 +96,3 @@ export class Topic {
     }
 }
 
-export class MetricBranch {
-}
-
-export class MetricTree {
-    constructor (payload) {
-        for (const m of payload.metrics) {
-            const path = m.name.split("/");
-            const name = path.pop();
-
-            let obj = this;
-            for (const p of path) {
-                obj = obj[p] ||= new MetricBranch();
-
-                if (!(obj instanceof MetricBranch)) {
-                    const branch = new MetricBranch();
-                    branch.$metric = obj;
-                    obj = branch;
-                }
-            }
-            obj[name] = m;
-        }
-    }
-}
